+++ Clients/TgDownloaderWinDesktop/Views/Pages/TgProxiesPage.xaml.cs
@@ -1,8 +1,6 @@
 ﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
-using TgDesktopUtils = TgDownloaderWinDesktop.Utils.TgDesktopUtils;
-
 namespace TgDownloaderWinDesktop.Views.Pages;
 
 /// <summary>
@@ -18,7 +16,6 @@ public partial class TgProxiesPage
 		TgDesktopUtils.TgProxiesVm.AddUpdateUi(TgEnumUpdateType.Window, UpdateMainWindow);
 		TgDesktopUtils.TgProxiesVm.AddUpdateUi(TgEnumUpdateType.Application, UpdateApplication);
 		TgDesktopUtils.TgProxiesVm.OnNavigatedTo();
-        TgDesktopUtils.TgProxiesVm.OnLoadProxiesFromStorageAsync().ConfigureAwait(false);
 		InitializeComponent();
 	}