+++ Core/TgStorage/Domain/Proxies/TgSqlTableProxyViewModel.cs
@@ -17,30 +17,28 @@ public sealed partial class TgSqlTableProxyViewModel : TgViewModelBase
         get => Proxy.Uid;
         set => Proxy = TgSqlTableProxyRepository.Instance.Get(value) ?? TgSqlTableProxyRepository.Instance.GetNew();
     }
-	public Action<TgSqlTableProxyViewModel> ConnectProxy { get; set; }
+	public Action<TgSqlTableProxyViewModel> ConnectClientByProxy { get; set; }
+	public Action<TgSqlTableProxyViewModel> DisconnectClient { get; set; }
 	public Action<TgSqlTableProxyViewModel> DeleteProxy { get; set; }
-	public Action LoadProxiesForClient { get; set; }
-	public Action LoadProxies { get; set; }
 
-	public string PrettyName => $"{Proxy.Type} | {Proxy.HostName} | {Proxy.Port} | {Proxy.UserName}";
+	public string PrettyName => $"{Proxy.Type} | {TgDataFormatUtils.GetFormatString(Proxy.HostName, 20)} | {Proxy.Port} | {Proxy.UserName}";
 
-	public TgSqlTableProxyViewModel(TgSqlTableProxyModel proxy, Action<TgSqlTableProxyViewModel> connectProxy,
-        Action<TgSqlTableProxyViewModel> deleteProxy, Action loadProxies, Action loadProxiesForClient)
+	public TgSqlTableProxyViewModel(TgSqlTableProxyModel proxy, 
+        Action<TgSqlTableProxyViewModel> connectClient, Action<TgSqlTableProxyViewModel> disconnectClient, 
+        Action<TgSqlTableProxyViewModel> deleteProxy)
 	{
 		Proxy = proxy;
-		ConnectProxy = connectProxy;
+		ConnectClientByProxy = connectClient;
+        DisconnectClient = disconnectClient;
         DeleteProxy = deleteProxy;
-        LoadProxies = loadProxies;
-		LoadProxiesForClient = loadProxiesForClient;
     }
 
 	public TgSqlTableProxyViewModel(TgSqlTableProxyModel proxy)
 	{
 		Proxy = proxy;
-		ConnectProxy = _ => { };
+		ConnectClientByProxy = _ => { };
+		DisconnectClient = _ => { };
         DeleteProxy = _ => { };
-        LoadProxies = () => { };
-        LoadProxiesForClient = () => { };
 	}
 
 	#endregion
@@ -50,12 +48,22 @@ public sealed partial class TgSqlTableProxyViewModel : TgViewModelBase
 	public override string ToString() => PrettyName;
 
 	[RelayCommand]
-	public async Task OnConnectProxyAsync()
+	public async Task OnConnectClientByProxyAsync()
 	{
 		await Task.Delay(TimeSpan.FromMilliseconds(1)).ConfigureAwait(false);
 		TgAsyncUtils.ExecAction(() =>
 		{
-			ConnectProxy(this);
+			ConnectClientByProxy(this);
+		});
+	}
+
+	[RelayCommand]
+	public async Task OnDisconnectClientAsync()
+	{
+		await Task.Delay(TimeSpan.FromMilliseconds(1)).ConfigureAwait(false);
+		TgAsyncUtils.ExecAction(() =>
+		{
+			DisconnectClient(this);
 		});
 	}
 
@@ -66,8 +74,6 @@ public sealed partial class TgSqlTableProxyViewModel : TgViewModelBase
 		TgAsyncUtils.ExecAction(() =>
 		{
             DeleteProxy(this);
-            LoadProxies();
-            LoadProxiesForClient();
         });
 	}